+++ src/mapping.js
@@ -1,4 +1,7 @@
 export const mapping = {
   contact: process.env.NODE_ENV === 'development' ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/contact' : 'https://contact-fjb46y5zza-uc.a.run.app',
   assistant: process.env.NODE_ENV === 'development' ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/assistant' : 'https://assistant-fjb46y5zza-uc.a.run.app',
+  quotation: process.env.NODE_ENV === 'development'
+    ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/quotation'
+    : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/quotation',
 }